requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.3.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
deepgram-sdk==3.*
frozenlist==1.4.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
gspread-dataframe==3.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
verboselogs==1.7
websockets==12.0
yarl==1.9.4
